server/pyproject.toml
[tool.poetry]
name = "family-ai-server"
version = "0.1.0"
description = "Family AI Companion FastAPI backend"
authors = ["Family AI Team <dev@familyai.local>"]
readme = "../README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
SQLAlchemy = "^2.0.28"
alembic = "^1.13.1"
psycopg = { extras = ["binary"], version = "^3.1.18" }
pgvector = "^0.2.5"
openai = "^1.14.2"
httpx = "^0.27.0"
tenacity = "^8.2.3"
bcrypt = "^4.1.2"
PyJWT = "^2.8.0"
tiktoken = "^0.6.0"
chromadb = "^0.4.24"
numpy = "<2.0"
python-dotenv = "^1.0.1"
boto3 = "^1.34.59"
loguru = "^0.7.2"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
ruff = "^0.3.5"
black = "^24.3.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.ruff.isort]
combine-as-imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["app/tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
